% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{AggregateArray}
\alias{AggregateArray}
\title{Performance aggregation across folds}
\usage{
AggregateArray(array, aggr.method = "median")
}
\arguments{
\item{array}{array; array of matrices with genes in rows and
methods in columns, where entries correspond to the the method's performance
for the specific gene.}

\item{aggr.method}{character; method used to aggregate performance results
from different folds. One of 'mean', 'median' (defaults to 'median').}
}
\value{
matrix; matrix with genes in rows and methods in columns, where
entries correspond to the performance of the method when imputing the given
gene.
}
\description{
\code{AggregateArray} aggregates the results of imputation
across different cross-validation folds into one value of imputation error
per gene and imputation method
}
\details{
The performances from different folds are aggregated into a single
value per gene, according to \code{aggr.method}.
}
\seealso{
\code{\link{ComputeMSE}}
}
