% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{AggregateMSEArray}
\alias{AggregateMSEArray}
\title{MSE aggregation across folds}
\usage{
AggregateMSEArray(MSE_array, aggr.method = "median")
}
\arguments{
\item{MSE_array}{array; array of matrices with genes in rows and methods in
columns, where entries correspond to the the method's MSE for the specific
gene.}

\item{aggr.method}{character; method used to aggregate MSE results from
different folds. One of 'mean', 'median' (defaults to 'median')}
}
\value{
matrix; matrix with genes in rows and methods in columns, where
entries correspond to the MSE of the method when imputing the given gene
}
\description{
\code{AggregateMSEArray} aggregates the results of imputation
across different cross-validation folds into one value of imputation error
per gene and imputation method
}
\details{
The MSE results from different folds are aggregated into a single
value per gene, according to \code{aggr.method}.
}
\seealso{
\code{\link{ComputeMSE}}
}
