% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{ArrangeData}
\alias{ArrangeData}
\title{Data trimming}
\usage{
\code{ArrangeData(data, network.path = NULL)}
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed (genes as rows
and samples as columns)}

\item{network.path}{character; path to .txt or .rds file with network
coefficients}
}
\value{
list; data matrix, network coefficients matrix and intercept for
genes common between the data matrix and the network
}
\description{
\code{ArrangeData} finds common genes to the network and
provided data and limits both datasets to these
}
