% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{CheckArguments}
\alias{CheckArguments}
\title{Argument check}
\usage{
CheckArguments(env)
}
\arguments{
\item{env}{environment where variables to test are located.}
}
\value{
collection of all messages.
}
\description{
\code{CheckArguments} tests for argument correctness.
}
