% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{Combine}
\alias{Combine}
\title{Combine imputation methods}
\usage{
Combine(data, imputed, method.choice, write = FALSE)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, already
normalized (genes as rows and samples as columns)}

\item{imputed}{list; list of matrices with imputation results for all
considered methods}

\item{method.choice}{named character; vector with the best performing method
per gene}

\item{write}{logical; should a file with the imputation results be written?}
}
\value{
matrix; imputation results combining the best performing method
per gene
}
\description{
Combine imputation methods
}
\details{
Combines imputation results from all methods according to training
results provided in \code{method.choice}
}
