% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{ComputeCor}
\alias{ComputeCor}
\title{Correlation computation}
\usage{
ComputeCor(real, masked, imputed)
}
\arguments{
\item{real}{matrix; original gene expression data, i.e. before masking
(genes as rows and samples as columns)}

\item{masked}{matrix, logical indicating which entries were masked
(genes as rows and samples as columns)}

\item{imputed}{list; list of matrices with imputation results for all
considered methods}
}
\value{
matrix; matrix with genes in rows and methods in columns, where
entries correspond to the the method's Pearson correlation for the specific
gene.
}
\description{
\code{ComputeCor} computes imputation performance based on
Pearson correlation
}
\details{
The imputed values are compared to the real ones for every masked
entry in \code{real}. The Pearson correlation is computed for all masked
entries per gene.
}
\seealso{
\code{\link{ComputeCorGenewise}}
}
