% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{ComputeCorGenewise}
\alias{ComputeCorGenewise}
\title{Computation of Pearson correlation per gene}
\usage{
ComputeCorGenewise(real, masked, imputed, baseline)
}
\arguments{
\item{real}{numeric; vector of original expression of a given gene (before
masking)}

\item{masked}{logical; vector indicating which entries were masked for a
given gene}

\item{imputed}{matrix; imputation results for a given imputation method}

\item{baseline}{logical; is this baseline imputation?}
}
\value{
Pearson correlation of all imputations indicated by \code{masked}
}
\description{
\code{ComputeCorGenewise} computes the Pearson correlation of
dropout imputation for a given gene.
}
