% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{ComputeMSE}
\alias{ComputeMSE}
\title{MSE computation}
\usage{
ComputeMSE(real, masked, imputed)
}
\arguments{
\item{real}{matrix; original gene expression data, i.e. before masking
(genes as rows and samples as columns)}

\item{masked}{matrix, logical indicating which entries were masked
(genes as rows and samples as columns)}

\item{imputed}{list; list of matrices with imputation results for all
considered methods}
}
\value{
matrix; matrix with genes in rows and methods in columns, where
entries correspond to the the method's MSE for the specific gene.
}
\description{
\code{ComputeMSE} determines the method for dropout
imputation based on performance on each gene in training data
}
\details{
The imputed values are compared to the real ones for every masked
entry in \code{real}. The Mean Squared Error is computed for all masked
entries per gene.
}
\seealso{
\code{\link{ComputeMSEGenewise}}
}
