% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{CreateArgCheck}
\alias{CreateArgCheck}
\title{Argument check}
\usage{
CreateArgCheck(missing = NULL, match = NULL, acceptable = NULL)
}
\arguments{
\item{missing}{named list; logical. Name corresponds to variable name, and
corresponding entry to whether it was missing from the function call.}

\item{match}{named list. Name corresponds to variable name, and corresponding
entry to its value.}

\item{acceptable}{named list. Name corresponds to variable name, and
corresponding entry to its acceptable values.}
}
\value{
argument check object.
}
\description{
\code{CreateArgCheck} creates tests for argument correctness.
}
