% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{CreateTrainData}
\alias{CreateTrainData}
\title{Preparation of training data for method evaluation}
\usage{
CreateTrainData(data, train.ratio = .7, train.only = TRUE, mask = .1,
write = FALSE)
}
\arguments{
\item{data}{matrix; raw counts (genes as rows and samples as columns)}

\item{train.ratio}{numeric; ratio of the samples to be used for training}

\item{train.only}{logical; if TRUE define only a training dataset, if
FALSE writes both training and validation sets (defaults to TRUE)}

\item{mask}{numeric; ratio of total non-zero samples to be masked per gene
(defaults to .1)}

\item{write}{logical; should the output be written to a file?}
}
\value{
list with resulting matrix after subsetting and after masking
}
\description{
\code{CreateTrainingData} selects a subset of cells to use as
training set and sets a portion (\code{mask}) of the non-zero entries in each
row of the subset to zero
}
