% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{DataCheck_Matrix}
\alias{DataCheck_Matrix}
\title{Data check (matrix)}
\usage{
DataCheck_Matrix(data)
}
\arguments{
\item{data}{data object to check}
}
\value{
data object with needed adjustments
}
\description{
\code{DataCheck_Matrix} tests for potential format and storage
issues with matrices. Helper function to ADImpute.
}
