% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{DataCheck_Network}
\alias{DataCheck_Network}
\title{Data check (network)}
\usage{
DataCheck_Network(network)
}
\arguments{
\item{network}{data object containing matrix coefficients}
}
\value{
network data object with needed adjustments
}
\description{
\code{DataCheck_Network} tests for potential format and storage
issues with the network coefficient matrix. Helper function to ADImpute.
}
