% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{DataCheck_SingleCellExperiment}
\alias{DataCheck_SingleCellExperiment}
\title{Data check (SingleCellExperiment)}
\usage{
DataCheck_SingleCellExperiment(sce, normalized = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment; data for normalization or imputation}

\item{normalized}{logical; is the data expected to be normalized?}
}
\value{
NULL object.
}
\description{
\code{DataCheck_SingleCellExperiment} tests for existence of the
appropriate assays in \code{sce}. Helper function to ADImpute.
}
