% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{DataCheck_TrLength}
\alias{DataCheck_TrLength}
\title{Data check (transcript length)}
\usage{
DataCheck_TrLength(trlength)
}
\arguments{
\item{trlength}{data object containing transcript length information}
}
\value{
transcript length object with needed adjustments
}
\description{
\code{DataCheck_TrLength} tests for potential format and
storage issues with the object encoding transcript length, for e.g. TPM
normalization. Helper function to ADImpute.
}
