% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{demo_net}
\alias{demo_net}
\title{Small regulatory network for example purposes}
\format{
matrix; subset of the Gene Regulatory Network installed along with
ADImpute.
}
\usage{
demo_net
}
\description{
Subset of the Gene Regulatory Network used by ADImpute's Network
imputation method.
}
\keyword{datasets}
