% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{demo_sce}
\alias{demo_sce}
\title{Small dataset for example purposes}
\format{
SingleCellExperiment; a subset of the Grun pancreas dataset, obtained
with the \code{scRNAseq} R package, to use in the vignette and examples.
}
\usage{
demo_sce
}
\description{
A small dataset to use on vignettes and examples (50 cells).
}
\references{
Grun D et al. (2016). De novo prediction of stem cell identity
using single-cell transcriptome data. Cell Stem Cell 19(2), 266-277.
}
\keyword{datasets}
