% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrap.R
\name{EvaluateMethods}
\alias{EvaluateMethods}
\title{Imputation method evaluation on training set}
\usage{
EvaluateMethods(data, training.ratio = .7, training.only = T,
mask.ratio = .1, split.seed = NULL, mask.seed = NULL, scale = 1,
pseudo.count = 1, labels = NULL, cell.clusters = 2, drop_thre = NULL,
type = "TPM", cores = 4, cluster.type = "SOCK", network.path = NULL,
transcript.length = NULL, drop.exclude = T, bulk = NULL, ...)
}
\arguments{
\item{data}{matrix; normalized counts, not logged (genes as rows and samples
as columns)}

\item{training.ratio}{numeric; ratio of samples to be used for training}

\item{training.only}{logical; if TRUE define only a training dataset, if
FALSE writes and returns both training and validation sets (defaults to T)}

\item{mask.ratio}{numeric; ratio of samples to be masked per gene}

\item{split.seed}{integer; optional seed (defaults to NULL, random selection
of samples to use for training)}

\item{mask.seed}{integer; optional seed (defaults to NULL, random selection
of samples to mask)}

\item{scale}{integer; scaling factor to divide all expression levels by
(defaults to 1)}

\item{pseudo.count}{integer; pseudo-count to be added to expression levels
to avoid log(0) (defaults to 1)}

\item{labels}{character; vector specifying the cell type of each column of
\code{data}}

\item{cell.clusters}{integer; number of cell subpopulations}

\item{drop_thre}{numeric; between 0 and 1 specifying the threshold to
determine dropout values}

\item{type}{A character specifying the type of values in the expression
matrix. Can be "count" or "TPM"}

\item{cores}{integer; number of cores used for paralell computation}

\item{cluster.type}{character; either "SOCK" or "MPI"}

\item{network.path}{character; path to .txt or .rds file with network
coefficients}

\item{transcript.length}{matrix with at least 2 columns: "hgnc_symbol" and
"transcript_length"}

\item{drop.exclude}{logical; should zeros be discarded for the calculation
of genewise average expression levels? (defaults to T)}

\item{bulk}{vector of reference bulk RNA-seq, if available (average across
samples)}

\item{...}{additional parameters to pass to network-based imputation}
}
\value{
character; best performing method in the training set for each gene
}
\description{
\code{EvaluateMethods} returns the best-performing imputation
method for each gene in the dataset
}
\details{
For each gene, a fraction (\code{mask.ratio}) of the quantified
expression values are set to zero and imputed according to 3 different
methods: scImpute, baseline (average gene expression across all cells) or a
network-based method. The imputation error is computed for each of the
values in the original dataset that was set to 0, for each method. The
method resulting in a lowest imputation error for each gene is chosen.
}
\seealso{
\code{\link{ImputeBaseline}},
\code{\link{ImputeDrImpute}},
\code{\link{ImputeNetwork}},
\code{\link{ImputeSAVER}},
\code{\link{ImputeScImpute}},
\code{\link{ImputeSCRABBLE}}
}
