% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeBaseline}
\alias{ImputeBaseline}
\title{Impute using average expression across all cells}
\usage{
ImputeBaseline(data, write = FALSE, ...)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, normalized
and log2-transformed (genes as rows and samples as columns)}

\item{write}{logical; should a file with the imputation results be written?}

\item{...}{additional arguments to \code{saveRDS}}
}
\value{
matrix; imputation results considering the average expression
values of genes
}
\description{
\code{ImputeBaseline} imputes dropouts using gene averages
across cells. Zero values are excluded from the mean computation.
}
