% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeDrImpute}
\alias{ImputeDrImpute}
\title{Use DrImpute}
\usage{
ImputeDrImpute(data, write = FALSE)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, normalized
and log2-transformed (genes as rows and samples as columns)}

\item{write}{logical; should a file with the imputation results be written?}
}
\value{
matrix; imputation results from DrImpute
}
\description{
\code{ImputeDrImpute} uses the DrImpute package for dropout
imputation
}
\seealso{
\code{\link[DrImpute]{DrImpute}}
}
