% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeNetwork}
\alias{ImputeNetwork}
\title{Network-based imputation}
\usage{
ImputeNetwork(data, network.coefficients = NULL, network.path = NULL,
cores = 4, type = "iteration", write.to.file = TRUE,
drop.exclude = TRUE, ...)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, normalized
and log2-transformed (genes as rows and samples as columns)}

\item{network.coefficients}{matrix; network coefficients. Please provide
either \code{network.coefficients} or \code{network.path}.}

\item{network.path}{character; path to .txt or .rds file with network
coefficients}

\item{cores}{integer; number of cores to use}

\item{type}{character; either "iteration", for an iterative solution, or
"pseudoinv", to use Moore-Penrose pseudo-inversion as a solution.}

\item{write.to.file}{logical; should a file with the imputation results be
written?}

\item{drop.exclude}{logical; should zeros be discarded for the calculation
of genewise average expression levels? (defaults to TRUE)}

\item{...}{additional arguments to \code{ImputeNetParallel}}
}
\value{
matrix; imputation results incorporating network information
}
\description{
Network-based imputation
}
\details{
Imputes dropouts using a gene regulatory network trained on external
data, as provided in \code{network.path}. Dropout expression values are
estimated from the expression of their predictor genes and the network
coefficients.
}
\seealso{
\code{\link{ImputeNetParallel}}
}
