% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{ImputeNonPredictiveDropouts}
\alias{ImputeNonPredictiveDropouts}
\title{Helper function to PseudoInverseSolution_percell}
\usage{
ImputeNonPredictiveDropouts(net, expr)
}
\arguments{
\item{net}{matrix, logical; network coefficients for all dropout (to be
imputed) genes that are predictive of the expression of other dropout genes}

\item{expr}{numeric; vector of gene expression for all genes in the cell at
hand}
}
\value{
vector; imputation results for the non-dropout-dependent genes
}
\description{
\code{ImputePredictiveDropouts} computes the non-dropout-
dependent solution of network imputation for each cell
}
