% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{ImputePredictiveDropouts}
\alias{ImputePredictiveDropouts}
\title{Helper function to PseudoInverseSolution_percell}
\usage{
ImputePredictiveDropouts(net, thr = 0.01, expr)
}
\arguments{
\item{net}{matrix, logical; network coefficients for all dropout (to be
imputed) genes that are predictive of the expression of other dropout genes}

\item{thr}{numeric; tolerance threshold to detect zero singular values}

\item{expr}{numeric; vector of gene expression for all genes in the cell at
hand}
}
\value{
vector; imputation results for the dropout-dependent genes
}
\description{
\code{ImputePredictiveDropouts} applies Moore-Penrose
pseudo-inversion to compute the dropout-dependent solution of network
imputation for each cell
}
