% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeSAVER}
\alias{ImputeSAVER}
\title{Use SAVER}
\usage{
ImputeSAVER(data, cores, try.mean = FALSE, write.to.file = TRUE)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, normalized
(genes as rows and samples as columns)}

\item{cores}{integer; number of cores to use}

\item{try.mean}{logical; whether to additionally use mean gene expression
as prediction}

\item{write.to.file}{logical; should a file with the imputation results be
written?}
}
\value{
matrix; imputation results from SAVER
}
\description{
\code{ImputeSAVER} uses the SAVER package for dropout
imputation
}
\seealso{
\code{\link[SAVER]{saver}}
}
