% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeScImpute}
\alias{ImputeScImpute}
\title{Use scImpute}
\usage{
\code{ImputeScImpute(count_path, infile, outfile = "rds", out_dir,
labeled, drop_thre, Kcluster, labels, ncores = 4, type = "TPM",
transcript.length.file = NULL, genelen = NULL)}
}
\arguments{
\item{count_path}{A character specifying the full path of the raw count
matrix}

\item{infile}{A character specifying the type of file storing the raw count
matrix; can be "csv", "txt", or "rds". The input file shoule have rows
representing genes and columns representing cells, with its first row as cell
names and first column as gene names}

\item{outfile}{A character specifying the type of file storing the imputed
count matrix; can be "csv", "txt", or "rds"}

\item{out_dir}{A character specifying the full path of the output directory,
which is used to store all intermdediate and final outputs}

\item{labeled}{A logical value indicating whether cell type information is
available. \code{labels} must be specified if \code{labeled = TRUE}}

\item{drop_thre}{A number between 0 and 1, specifying the threshold to
determine dropout values}

\item{Kcluster}{An integer specifying the number of cell subpopulations.
This parameter can be determined based on prior knowledge or clustering of
raw data. Kcluster is used to determine the candidate neighbors of each cell}

\item{labels}{A character vector specifying the cell type of each column in
the raw count matrix. Only needed when \code{labeled = TRUE}. Each cell type
should have at least two cells for imputation}

\item{ncores}{A integer specifying the number of cores used for parallel
computation}

\item{type}{A character specifying the type of values in the expression
matrix. Can be "count" (default) or "TPM"}

\item{genelen}{An integer vector giving the length of each gene. Order must
match the gene orders in the expression matrix. genelen must be specified if
\code{type = "count"}}
}
\value{
matrix; imputation results from scImpute
}
\description{
\code{ImputeScImpute} uses the scImpute package for dropout
imputation
}
\seealso{
\code{\link{scImpute::scimpute}}
}
