% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeSCRABBLE}
\alias{ImputeSCRABBLE}
\title{Use SCRABBLE}
\usage{
\code{ImputeSCRABBLE(data, bulk = NULL, write.to.file = T)}
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, normalized
(genes as rows and samples as columns)}

\item{bulk}{vector of reference bulk RNA-seq, if available (average across
samples)}

\item{write.to.file}{logical; should a file with the imputation results be
written?}
}
\value{
matrix; imputation results from SCRABBLE
}
\description{
\code{ImputeSCRABBLE} uses the SCRABBLE package for dropout
imputation
}
\seealso{
\code{\link{SCRABBLE::scrabble}}
}
