% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{network.coefficients}
\alias{network.coefficients}
\title{Transcriptome wide gene regulatory network}
\format{
dgCMatrix
}
\usage{
network.coefficients
}
\description{
Gene Regulatory Network used by ADImpute's Network imputation
method. First column, \code{O}, corresponds to the intercept of a gene-
-specific predicion model. The remaining rows and columns correspond to the
adjacency matrix of the inferred network, where rows are target genes and
columns are predictors. Genes are identified by their hgnc_symbol.
}
\keyword{datasets}
