% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{NormalizeRPM}
\alias{NormalizeRPM}
\title{RPM normalization}
\usage{
NormalizeRPM(data, sce = NULL, log = FALSE, scale = 1,
pseudo.count = 1)
}
\arguments{
\item{data}{matrix; raw data (genes as rows and samples as columns)}

\item{sce}{SingleCellExperiment; raw data}

\item{log}{logical; log RPMs?}

\item{scale}{integer; scale factor to divide RPMs by}

\item{pseudo.count}{numeric; if \code{log = TRUE}, value to add to RPMs in
order to avoid taking \code{log(0)}}
}
\value{
matrix; library size normalized data
}
\description{
\code{NormalizeRPM} performs RPM normalization, with possibility
to log the result
}
\examples{
demo <- NormalizeRPM(ADImpute::demo_data)

}
