% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{PseudoInverseSolution_percell}
\alias{PseudoInverseSolution_percell}
\title{Network-based parallel imputation - Moore-Penrose pseudoinversion}
\usage{
PseudoInverseSolution_percell(expr, net, drop_ind, thr = 0.01)
}
\arguments{
\item{expr}{numeric; expression vector for cell at hand}

\item{net}{matrix; network coefficients}

\item{drop_ind}{logical; dropout entries in the cell at hand}

\item{thr}{numeric; tolerance threshold to detect zero singular values}
}
\value{
matrix; imputation results incorporating network information
}
\description{
\code{PseudoInverseSolution_percell} applies Moore-Penrose
pseudo-inversion to compute the solution of network imputation for each
cell
}
