% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Write.R
\name{ReadData}
\alias{ReadData}
\title{Data read}
\usage{
\code{ReadData(path, ...)}
}
\arguments{
\item{path}{character; path to input file}
}
\value{
data.frame; raw counts (genes as rows and samples as columns)
}
\description{
\code{ReadData} reads data from raw input file (.txt or .csv)
}
