% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{ReadNetwork}
\alias{ReadNetwork}
\title{Network loading}
\usage{
ReadNetwork(network.path)
}
\arguments{
\item{network.path}{character; path to .txt, .rds or .zip file with network
coefficients}
}
\value{
matrix with network coefficients and intercept in the first column
}
\description{
\code{ReadNetwork} loads the matrix of network coefficients
}
