% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{ReturnOut}
\alias{ReturnOut}
\title{Wrapper for return of Impute()}
\usage{
ReturnOut(result, sce)
}
\arguments{
\item{result}{list; imputation result}

\item{sce}{SingleCellExperiment; a SingleCellExperiment object if available;
NULL otherwise}
}
\value{
imputation results. A SingleCellExperiment if \code{!is.null(sce)},
or a list with imputed results in matrix format otherwise.
}
\description{
\code{ReturnOut} Adjusts the output of \code{Impute} to a list
of matrices or a SingleCellExperiment object. Helper function to ADImpute.
}
