% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{transcript_length}
\alias{transcript_length}
\title{Table for transcript length calculations}
\format{
A data.frame with 2 columns:
\describe{
\item{hgnc_symbol}{Gene symbol identifier}
\item{transcript length}{Length of transcript}
}
}
\usage{
transcript_length
}
\description{
A data.frame to be used for transcript length computations.
May be necessary upon TPM normalization, or as input to \code{scImpute}.
All data was retrieved from \code{biomaRt}.
}
\keyword{datasets}
