% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Write.R
\name{WriteCSV}
\alias{WriteCSV}
\title{Write csv file}
\usage{
WriteCSV(object, file)
}
\arguments{
\item{object}{R object to write}

\item{file}{character; path to output file}
}
\description{
\code{WriteCSV} writes data to a comma-delimited output file
}
