% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Write.R
\name{WriteTXT}
\alias{WriteTXT}
\title{Write txt file}
\usage{
WriteTXT(object, file)
}
\arguments{
\item{object}{R object to write}

\item{file}{character; path to output file}
}
\value{
Returns NULL
}
\description{
\code{WriteTXT} writes data to a tab-delimited output file
}
\examples{
WriteTXT(demo_data_50cells, file = "demo.txt")

}
