% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{ChooseMethod}
\alias{ChooseMethod}
\title{Method choice per gene}
\usage{
ChooseMethod(MSE, write.to.file = TRUE)
}
\arguments{
\item{MSE}{matrix; matrix with genes in rows and methods in columns, where
entries correspond to the the method's MSE for the specific gene.}

\item{write.to.file}{logical; should the output be written to a file?}
}
\value{
character; best performing method in the training set for each gene
}
\description{
\code{ChooseMethod} determines the method for dropout
imputation based on performance on each gene in training data
}
\details{
The method with the best performance (lowest MSE) is chosen for each
gene.
}
\seealso{
\code{\link{ComputeMSE}}
}
