% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{ComputeMSEGenewise}
\alias{ComputeMSEGenewise}
\title{Computation of MSE per gene}
\usage{
\code{ComputeMSEGenewise(real, masked, scimputed, baseline, network)}
}
\arguments{
\item{real}{numeric; vector of original expression of a given gene (before
masking)}

\item{masked}{logical; vector indicating which entries were masked for a
given gene}

\item{scimputed}{numeric; vector of results of scImpute for a given gene}

\item{baseline}{numeric; vector of results of baseline imputation for a
given gene}

\item{network}{numeric; vector of results of network imputation for a given
gene}
}
\value{
MSE of all imputations for a given gene, for each of the three
methods: scImpute, average expression (baseline) and network.
}
\description{
\code{ComputeMSEGenewise} computes the MSE of dropout
imputation for a given gene.
}
