% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Dataset for example purposes}
\format{
A matrix of 100 random cells sampled from a publicly
available dataset on hESC differention by Chu et al.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75748}{GEO accession GSE75748}
}
\usage{
demo_data
}
\description{
A dataset to use on vignettes and examples (100 cells).
}
\references{
Chu, L., Leng, N., Zhang, J. et al. Single-cell RNA-seq reveals
novel regulators of human embryonic stem cell differentiation to definitive
endoderm. Genome Biol 17, 173 (2016).
\href{https://doi.org/10.1186/s13059-016-1033-x}{DOI}
}
\keyword{datasets}
