% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrap.R
\name{EvaluateMethods}
\alias{EvaluateMethods}
\title{Imputation method evaluation on training set}
\usage{
\code{EvaluateMethods(data, training.ratio = .7, training.only = T,
mask.ratio = .1, split.seed = NULL, mask.seed = NULL, log = F, scale = 1,
pseudo.count = 1, labels = NULL, cell.clusters = NULL, drop_thre = NULL,
cores = 4, cluster.type = "SOCK", network.path = "../network/network.rds",
...)}
}
\arguments{
\item{data}{matrix; normalized counts, not logged (genes as rows and samples
as columns)}

\item{training.ratio}{numeric; ratio of samples to be used for training}

\item{training.only}{logical; if TRUE define only a training dataset, if
FALSE writes and returns both training and validation sets (defaults to T)}

\item{mask.ratio}{numeric; ratio of samples to be masked per gene}

\item{split.seed}{integer; optional seed (defaults to NULL, random selection
of samples to use for training)}

\item{mask.seed}{integer; optional seed (defaults to NULL, random selection
of samples to mask)}

\item{scale}{integer; scaling factor to divide all expression levels by
(defaults to 1)}

\item{pseudo.count}{integer; pseudo-count to be added to expression levels
to avoid log(0) (defaults to 1)}

\item{labels}{character; vector specifying the cell type of each column of
\code{data}}

\item{cell.clusters}{integer; number of cell subpopulations}

\item{drop_thre}{numeric; between 0 and 1 specifying the threshold to
determine dropout values}

\item{cores}{integer; number of cores used for paralell computation}

\item{cluster.type}{character; either "SOCK" or "MPI"}

\item{network.path}{character; path to .rds file containing network
coefficients and intercept}
}
\value{
character; best performing method in the training set for each gene
}
\description{
\code{EvaluateMethods} returns the best-performing imputation
method for each gene in the dataset
}
\details{
For each gene, a fraction (\code{mask.ratio}) of the quantified
expression values are set to zero and imputed according to 3 different
methods: scImpute, baseline (average gene expression across all cells) or a
network-based method. The imputation error is computed for each of the
values in the original dataset that was set to 0, for each method. The
method resulting in a lowest imputation error for each gene is chosen.
}
\seealso{
\code{\link{ImputeBaseline}},
\code{\link{ImputeNetwork}},
\code{\link{ImputeScImpute}},
\code{\link{scimpute}}
}
