% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeBaseline}
\alias{ImputeBaseline}
\title{Impute using average expression across all cells}
\usage{
ImputeBaseline(data, write.to.file = T, ...)
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed (genes as rows
and samples as columns)}

\item{write.to.file}{logical; should a file with the imputation results be
written?}
}
\value{
matrix; imputation results considering the average expression
values of genes
}
\description{
\code{ImputeBaseline} imputes dropouts using gene averages
across cells
}
