% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{ImputeNetParallel}
\alias{ImputeNetParallel}
\title{Network-based parallel imputation}
\usage{
\code{ImputeNetParallel(dropout.matrix, arranged, cores = 4,
cluster.type = "SOCK", max.iter = 20)}
}
\arguments{
\item{dropout.matrix}{matrix, logical; dropout entries in the data matrix
(genes as rows and samples as columns)}

\item{arranged}{list; output of \code{\link{ArrangeData}}}

\item{cores}{integer; number of cores used for paralell computation}

\item{cluster.type}{character; either "SOCK" or "MPI"}

\item{max.iter}{numeric; maximum number of iterations for network
imputation. Set to -1 to remove limit (not recommended)}
}
\value{
matrix; imputation results incorporating network information
}
\description{
\code{ImputeNetParallel} implements network-based imputation
in parallel
}
