% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{ImputeNetwork}
\alias{ImputeNetwork}
\title{Network-based imputation}
\usage{
\code{ImputeNetwork(data, network.path = NULL, cores = 4,
cluster.type = "SOCK", write.to.file = T, ...)}
}
\arguments{
\item{data}{matrix with entries equal to zero to be imputed, already
normalized (genes as rows and samples as columns)}

\item{network.path}{character; path to .rds file containing network
coefficients and intercept}

\item{cores}{integer; number of cores to use (parallel computation if cores
> 1)}

\item{write.to.file}{logical; should a file with the imputation results be
written?}
}
\value{
matrix; imputation results incorporating network information
}
\details{
Imputes dropouts using a gene regulatory network trained on external
data, as provided in \code{network.path}. Dropout expression values are
estimated from the expression of their predictor genes and the network
coefficients.
}
\seealso{
\code{\link{ImputeNetParallel}}
}
