% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{MaskData}
\alias{MaskData}
\title{Masking of entries for performance evaluation}
\usage{
\code{MaskData(data, write.to.file = T, filename, mask = .1,
seed = NULL)}
}
\arguments{
\item{data}{matrix; raw counts (genes as rows and samples as columns)}

\item{filename}{character; name of output file to write masked data to}

\item{mask}{numeric; ratio of total non-zero samples to be masked per gene
(defaults to .1)}

\item{seed}{integer; optional seed (defaults to NULL, random selection of
samples to mask)}
}
\value{
matrix containing masked raw counts (genes as rows and samples as
columns)
}
\description{
\code{MaskData} sets a portion (\code{mask}) of the non-zero
entries of each row of \code{data} to zero
}
\details{
Sets a portion (\code{mask}) of the non-zero entries of each row of
\code{data} to zero. For reproducible results, provide seed.
\code{seed} is summed to row number so that every gene has a different seed
(different samples are masked for each gene).
Result is written to \code{filename}.
}
