% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenewiseMethod.R
\name{MaskerPerGene}
\alias{MaskerPerGene}
\title{Helper mask function}
\usage{
MaskerPerGene(x, rowmask)
}
\arguments{
\item{x}{numeric; data to mask}

\item{rowmask}{numeric; number of samples to be masked per gene}
}
\value{
numeric containing masked raw counts
}
\description{
Helper mask function, per feature.
}
