% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{NormalizeRPM}
\alias{NormalizeRPM}
\title{RPM normalization}
\usage{
\code{NormalizeTPM(data, log = F, scale = 1, pseudo.count = 1)}
}
\arguments{
\item{data}{matrix; raw data (genes as rows and samples as columns)}

\item{log}{logical; log RPMs?}

\item{scale}{integer; scale factor to divide RPMs by}

\item{pseudocount}{numeric; if \code{log = T}, value to add to RPMs in order
to avoid taking \code{log(0)}}
}
\value{
matrix; library size normalized data
}
\description{
\code{NormalizeRPM} performs RPM normalization, with possibility
to log the result
}
