% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFunctions.R
\name{PseudoInverseSolution_percell}
\alias{PseudoInverseSolution_percell}
\title{Network-based parallel imputation - Moore-Penrose pseudoinversion}
\usage{
PseudoInverseSolution_percell(cell, arranged, dropout_mat,
thr = 0.01)
}
\arguments{
\item{cell}{numeric; index of the column corresponding to current cell}

\item{arranged}{list; output of \code{\link{ArrangeData}}}

\item{dropout_mat}{matrix, logical; dropout entries in the data matrix
(genes as rows and samples as columns)}

\item{thr}{numeric; tolerance threshold to detect zero singular values}
}
\value{
matrix; imputation results incorporating network information
}
\description{
\code{PseudoInverseSolution_percell} applies Moore-Penrose
pseudo-inversion to compute the solution of network imputation for each
cell
}
